transcendence/data_script/requirements.txt
pandas
sqlalchemy
psycopg2-binary
matplotlib
celery
redis
fpdf2